+++ src/gdt.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use x86_64::VirtAddr;
 use x86_64::structures::gdt::{GlobalDescriptorTable, Descriptor, SegmentSelector};
 use x86_64::structures::tss::TaskStateSegment;