+++ src/lib.rs
@@ -1,22 +1,31 @@
 #![no_std]
 #![cfg_attr(test, no_main)]
+#![feature(alloc_error_handler)]
 #![feature(custom_test_frameworks)]
 #![feature(abi_x86_interrupt)]
 #![test_runner(crate::testexec)]
 #![reexport_test_harness_main = "testmain"]
 
 
-//extern crate alloc;
+extern crate alloc;
 extern crate core;
 
 use core::ops::Deref;
 use core::panic::PanicInfo;
 
+pub mod allocator;
 pub mod gdt;
 pub mod intr;
 pub mod mem;
 pub mod ser;
 pub mod vgabuff;
+
+
+#[alloc_error_handler]
+fn alloc_error_handler(layout: alloc::alloc::Layout) -> !
+{
+	panic!("[ERR] ALLOC ERR: {:?}", layout)
+}
 
 pub fn init()
 {