+++ src/main.rs
@@ -6,8 +6,11 @@
 #![allow(dead_code)]
 #![allow(deprecated)]
 #![allow(unused_features)]
+#![allow(unused_variables)]
 
+extern crate alloc;
 
+use alloc::{boxed::Box, vec, vec::Vec, rc::Rc};
 use bootloader::{BootInfo, entry_point};
 use core::panic::PanicInfo;
 use libertyos_kernel::println;
@@ -18,6 +21,7 @@ entry_point!(kernel_main);
 fn kernel_main(bootinfo: &'static BootInfo) -> !
 {
 	use libertyos_kernel::mem::{self, BootInfoFrameAllocator};
+	use libertyos_kernel::allocator;
 	use x86_64::{structures::paging::Page, VirtAddr};
 
 	println!("LIBERTY-OS");
@@ -42,6 +46,25 @@ fn kernel_main(bootinfo: &'static BootInfo) -> !
 	{
 		page_ptr.offset(400).write_volatile(0x_f021_f077_f065_f04e)
 	};
+
+	allocator::init_heap(&mut mapper, &mut framealloc)
+		.expect("[ERROR] FAILED TO INITIALIZE HEAP");
+
+	let heap_value = Box::new(41);
+	println!("[INFO] HEAP_VALUE AT {:p}", heap_value);
+
+	let mut vec = Vec::new();
+	for i in 0..500
+	{
+		vec.push(i);
+	}
+	println!("[INFO] VEC AT {:p}", vec.as_slice());
+
+	let refcounted = Rc::new(vec![1, 2, 3]);
+	let clonedref = refcounted.clone();
+	println!("[INFO] CURRENT REFERENCE COUNT: {}", Rc::strong_count(&clonedref));
+	core::mem::drop(refcounted);
+	println!("[INFO] CURRENT REFERENCE COUNT: {}", Rc::strong_count(&clonedref));
 
 	#[cfg(test)]
 	testexec();